split index to many for guard to run properly
