+++ Guardfile
@@ -3,3 +3,5 @@ guard :rspec, cmd: 'rspec' do
   watch(%r{^lib/daru/(.+)\.rb$})     { |m| "spec/#{m[1]}_spec.rb" }
   watch('spec/spec_helper.rb')  { "spec" }
 end
+
+# TODO: Run Rubocop in the background
